fix(windmill): bump windmill to the scoped package
use `@windmill/react-ui` in place of `windmill-react-ui`

BREAKING CHANGE: Changed every occurrence of `windmill-react-ui` with the scoped
`@windmill/react-ui`. You should see no difference updating and you should update, as the other
project is now deprecated and will no longer be developed. This was needed because the `windmill`
org was in dispute and npm solved it today.